[docs] Improve the SvgIcon documentation (#18560)
