+++ docs/src/pages/components/icons/FontAwesome.js
@@ -1,20 +1,13 @@
 import React from 'react';
-import clsx from 'clsx';
 import { loadCSS } from 'fg-loadcss';
 import { makeStyles } from '@material-ui/core/styles';
-import { red } from '@material-ui/core/colors';
+import { green } from '@material-ui/core/colors';
 import Icon from '@material-ui/core/Icon';
 
 const useStyles = makeStyles(theme => ({
   root: {
     '& > .fa': {
       margin: theme.spacing(2),
-    },
-  },
-  iconHover: {
-    margin: theme.spacing(2),
-    '&:hover': {
-      color: red[800],
     },
   },
 }));
@@ -34,13 +27,9 @@ export default function FontAwesome() {
       <Icon className="fa fa-plus-circle" />
       <Icon className="fa fa-plus-circle" color="primary" />
       <Icon className="fa fa-plus-circle" color="secondary" />
-      <Icon className="fa fa-plus-circle" color="action" />
-      <Icon
-        className={clsx(classes.iconHover, 'fa fa-plus-circle')}
-        color="error"
-        style={{ fontSize: 30 }}
-      />
-      <Icon className="fa fa-plus-circle" color="disabled" fontSize="large" />
+      <Icon className="fa fa-plus-circle" style={{ color: green[500] }} />
+      <Icon className="fa fa-plus-circle" fontSize="small" />
+      <Icon className="fa fa-plus-circle" style={{ fontSize: 30 }} />
     </div>
   );
 }
