+++ docs/src/pages/components/icons/Icons.js
@@ -1,17 +1,12 @@
 import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
-import { red } from '@material-ui/core/colors';
+import { green } from '@material-ui/core/colors';
 import Icon from '@material-ui/core/Icon';
 
 const useStyles = makeStyles(theme => ({
   root: {
     '& > span': {
       margin: theme.spacing(2),
-    },
-  },
-  iconHover: {
-    '&:hover': {
-      color: red[800],
     },
   },
 }));
@@ -24,13 +19,9 @@ export default function Icons() {
       <Icon>add_circle</Icon>
       <Icon color="primary">add_circle</Icon>
       <Icon color="secondary">add_circle</Icon>
-      <Icon color="action">add_circle</Icon>
-      <Icon className={classes.iconHover} color="error" style={{ fontSize: 30 }}>
-        add_circle
-      </Icon>
-      <Icon color="disabled" fontSize="large">
-        add_circle
-      </Icon>
+      <Icon style={{ color: green[500] }}>add_circle</Icon>
+      <Icon fontSize="small">add_circle</Icon>
+      <Icon style={{ fontSize: 30 }}>add_circle</Icon>
     </div>
   );
 }
