+++ docs/src/pages/components/icons/SvgIconsSize.tsx
@@ -1,0 +1,34 @@
+import React from 'react';
+import { makeStyles, createStyles, Theme } from '@material-ui/core/styles';
+import SvgIcon, { SvgIconProps } from '@material-ui/core/SvgIcon';
+
+const useStyles = makeStyles((theme: Theme) =>
+  createStyles({
+    root: {
+      '& > svg': {
+        margin: theme.spacing(2),
+      },
+    },
+  }),
+);
+
+function HomeIcon(props: SvgIconProps) {
+  return (
+    <SvgIcon {...props}>
+      <path d="M10 20v-6h4v6h5v-8h3L12 3 2 12h3v8z" />
+    </SvgIcon>
+  );
+}
+
+export default function SvgIconsSize() {
+  const classes = useStyles();
+
+  return (
+    <div className={classes.root}>
+      <HomeIcon fontSize="small" />
+      <HomeIcon />
+      <HomeIcon fontSize="large" />
+      <HomeIcon style={{ fontSize: 40 }} />
+    </div>
+  );
+}
